pasa el timer a otro file nuevo
